Devel (#2)
* Trivial change

* removed the memory class

* added display handling for local control og ATU

* Update display_handling.py

Turning this into a class

* Update display_handling.py

buttons are now handled by the main code.

* Added Interrupt control from display switches

* Added some ToDos and isr for buttons

* Not using pimoroni button, use pin instead

* print network SSIDs to screen

* display list of SSIDs

* fix wifi connect bug

if there are less than 0 wifi ssids, then wait and try again. If there are 1-6 the display them

* more hardware definitions

* fix bug number of networks

* sort the SSID results

* fix zero network display bug

* Server testing

* onboard LED. Test pin on boot

Onboard LED on when logged in to WiFi network as per Phils request
Test button A on bootup to exit main. Useful for upgrading